Use `dsect` to automatically allocate zp addrs
A 'dummy section' has its own .org offset, but doesn't change
the .org of the 'real' section that it is included in.

I kept the addresses identical to my hand-allocated ones from
before, to ensure it generated identical machine code. If I
ever run out of space on the zero page, I can remove these
.org directives and 'defragment' the zero page.